requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
Pillow>=10.3.0
numpy>=1.26.0

